move tests to correct files
