+++ winterfell_adapter/src/examples/cubic.rs
@@ -69,3 +69,52 @@ pub fn build_trace(sequence_length: usize) -> TraceTable<Felt> {
     }
     TraceTable::init(vec![column])
 }
+
+#[cfg(test)]
+mod tests {
+    use miden_core::Felt;
+    use stark_platinum_prover::{
+        proof::options::ProofOptions, prover::IsStarkProver, verifier::IsStarkVerifier,
+    };
+    use winter_air::TraceInfo;
+    use winter_prover::{Trace, TraceTable};
+
+    use crate::{
+        adapter::{
+            air::AirAdapter, public_inputs::AirAdapterPublicInputs, Prover, Transcript, Verifier,
+        },
+        examples::cubic::{self, Cubic},
+    };
+
+    #[test]
+    fn prove_and_verify_a_winterfell_cubic_air() {
+        let lambda_proof_options = ProofOptions::default_test_options();
+        let winter_trace = cubic::build_trace(16);
+        let trace = AirAdapter::<Cubic, TraceTable<_>, Felt, ()>::convert_winterfell_trace_table(
+            winter_trace.main_segment().clone(),
+        );
+        let pub_inputs = AirAdapterPublicInputs {
+            winterfell_public_inputs: *trace.columns()[0][15].value(),
+            transition_exemptions: vec![1],
+            transition_offsets: vec![0, 1],
+            trace_info: TraceInfo::new(1, 16),
+            metadata: (),
+        };
+
+        let proof = Prover::prove::<AirAdapter<Cubic, TraceTable<_>, Felt, _>>(
+            &trace,
+            &pub_inputs,
+            &lambda_proof_options,
+            Transcript::new(&[]),
+        )
+        .unwrap();
+        assert!(
+            Verifier::verify::<AirAdapter<Cubic, TraceTable<_>, Felt, _>>(
+                &proof,
+                &pub_inputs,
+                &lambda_proof_options,
+                Transcript::new(&[]),
+            )
+        );
+    }
+}
